Adds example using FastLED library to flash LEDS
This compiles and installs on "Heltec Wifi LoRa 32 (v3)" board, and flashes a
LED on pin 7, and changes the color of a NeoPixel Led (WS28182B) attached to min
6.

The current 'fastled' library in ESPhome does not compile correctly for the
ESP32-S3 chip. It will compile and run with a patch for an ESP32 based
board (Heltec Wifi LoRa 32 (V2)).

Base Arduino library used:
- FastLED/FastLED https://github.com/FastLED/FastLED/wiki/Controlling-leds

Installed library with ESPhome
- fastled/FastLED@3.5.0

FastLED Library successfully used with ESP32 with patch (Heltec Wifi LoRa
32(V2)), using details for a fix here:
https://github.com/esphome/issues/issues/4079

external_components:
  - source: github://pr#4257
    components: [fastled]
    refresh: always

light:
  - platform: fastled_clockless
    chipset: WS2812B
    pin: GPIO13
    num_leds: 1
    rgb_order: GRB
    name: "RGB Light"