lora: Initial commit prior to migration to RabioLib
RadioLib is the LoRa library being used by Meshtastic for the LoRa radio.
The current library being used is 'SX126x-Arduino', but this library does not
allow the syncword to be set.

The aim is to port the lora code to use RadioLib so that it can
interoperate(receive) with Meshtastic.